v 0.2.5
* Check if project is already installed.
* Check if some folders and file already exists before creation.
* More plugins in default sublime project for WordPress.

+++ inc/settings-all.sh
@@ -1,12 +1,18 @@
 #!/bin/bash
 
 # Folders
-mkdir "${BASEDIR}logs";
-mkdir "${BASEDIR}backups";
+if [[ ! -d "${BASEDIR}logs" ]];then
+    mkdir "${BASEDIR}logs";
+fi;
+if [[ ! -d "${BASEDIR}backups" ]];then
+    mkdir "${BASEDIR}backups";
+fi;
 
 # WordPress
 if [[ "${_INSTALL_CMS}" == 'wordpress' ]];then
     # WP-CLI
-    curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+    if [[ ! -f "${BASEDIR}wp-cli.phar" ]];then
+        curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+    fi;
 fi;
