+++ inc/settings-local.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
 
 # Wakatime
-echo "${_PROJECT_NAME}" > "${BASEDIR}.wakatime";
+if [[ ! -f  "${BASEDIR}.wakatime" ]];then
+    echo "${_PROJECT_NAME}" > "${BASEDIR}.wakatime";
+fi;
 
 # WordPress
 if [[ "${_INSTALL_CMS}" == 'wordpress' ]];then
     # Sublime Project
-    cp "${SCRIPTDIR}/tpl/sublime-project__wordpress.txt" "${BASEDIR}${_PROJECT_NAME}.sublime-project";
+    if [[ ! -f "${BASEDIR}${_PROJECT_NAME}.sublime-project" ]];then
+        cp "${SCRIPTDIR}/tpl/sublime-project__wordpress.txt" "${BASEDIR}${_PROJECT_NAME}.sublime-project";
+    fi;
 fi;
